Backend/requirements.txt
flask
flask-pymongo
flask-cors
python-dotenv
pyjwt
dnspython